requirements.txt
fastapi
uvicorn
google-cloud_ndb
python-dotenv
firebase-admin